move getImageProperties, getBlob to utils/canvas and use existing functions if possible
