+++ src/layer/const.ts
@@ -122,6 +122,13 @@ export const MimeTypes: Record<string, MimeType> = {
   PNG: 'image/png',
 };
 
+export type ImageProperties = {
+  rgba: Uint8ClampedArray;
+  width: number;
+  height: number;
+  format: MimeType;
+};
+
 export const SH_SERVICE_HOSTNAMES_V1_OR_V2: string[] = ['https://eocloud.sentinel-hub.com/'];
 
 export const SH_SERVICE_HOSTNAMES_V3: string[] = [
