+++ src/mapDataManipulation/mapDataManipulationUtils.ts
@@ -1,57 +1,4 @@
-import { Effects, ColorRange, ImageProperties } from './const';
-
-export async function getImageData(originalBlob: Blob): Promise<ImageProperties> {
-  let imgObjectUrl: any;
-  const imgCanvas = document.createElement('canvas');
-  try {
-    const imgCtx = imgCanvas.getContext('2d');
-    imgObjectUrl = window.URL.createObjectURL(originalBlob);
-    const img: any = await new Promise((resolve, reject) => {
-      const img = new Image();
-      img.onload = () => resolve(img);
-      img.onerror = reject;
-      img.src = imgObjectUrl;
-    });
-
-    imgCanvas.width = img.width;
-    imgCanvas.height = img.height;
-    imgCtx.drawImage(img, 0, 0);
-    const imgData = imgCtx.getImageData(0, 0, img.width, img.height).data;
-
-    return { rgba: imgData, width: img.width, height: img.height, format: originalBlob.type };
-  } catch (e) {
-    console.error(e);
-    throw new Error(e);
-  } finally {
-    imgCanvas.remove();
-    if (imgObjectUrl) {
-      window.URL.revokeObjectURL(imgObjectUrl);
-    }
-  }
-}
-
-export async function getBlob(imageProperties: ImageProperties): Promise<Blob> {
-  const { rgba, width, height, format } = imageProperties;
-  const imgCanvas = document.createElement('canvas');
-  try {
-    imgCanvas.width = width;
-    imgCanvas.height = height;
-    const imgCtx = imgCanvas.getContext('2d');
-    const newImg = new ImageData(rgba, width, height);
-    imgCtx.putImageData(newImg, 0, 0);
-    const blob: Blob = await new Promise(resolve => {
-      imgCanvas.toBlob(blob => {
-        resolve(blob);
-      }, format);
-    });
-    return blob;
-  } catch (e) {
-    console.error(e);
-    throw new Error(e);
-  } finally {
-    imgCanvas.remove();
-  }
-}
+import { Effects, ColorRange } from './const';
 
 // from one range to another
 // f(x) = c + ((d - c) / (b - a)) * (x - a)
