+++ src/utils/canvas.ts
@@ -1,4 +1,4 @@
-import { MimeType } from '../layer/const';
+import { MimeType, ImageProperties } from '../layer/const';
 
 export async function drawBlobOnCanvas(
   ctx: CanvasRenderingContext2D,
@@ -10,7 +10,7 @@ export async function drawBlobOnCanvas(
   try {
     // wait until objectUrl is drawn on the image, so you can safely draw img on canvas:
     const imgDrawn: HTMLImageElement = await new Promise((resolve, reject) => {
-      const img = document.createElement('img');
+      const img = new Image();
       img.onload = () => resolve(img);
       img.onerror = reject;
       img.src = objectURL;
@@ -23,4 +23,55 @@ export async function drawBlobOnCanvas(
 
 export async function canvasToBlob(canvas: HTMLCanvasElement, mimeFormat: MimeType): Promise<Blob> {
   return await new Promise(resolve => canvas.toBlob(resolve, mimeFormat));
+}
+
+export async function getImageProperties(originalBlob: Blob): Promise<ImageProperties> {
+  let imgObjectUrl: any;
+  const imgCanvas = document.createElement('canvas');
+  try {
+    const imgCtx = imgCanvas.getContext('2d');
+    imgObjectUrl = window.URL.createObjectURL(originalBlob);
+    const img: any = await new Promise((resolve, reject) => {
+      const img = new Image();
+      img.onload = () => resolve(img);
+      img.onerror = reject;
+      img.src = imgObjectUrl;
+    });
+    imgCanvas.width = img.width;
+    imgCanvas.height = img.height;
+    imgCtx.drawImage(img, 0, 0);
+    const imgData = imgCtx.getImageData(0, 0, img.width, img.height).data;
+
+    const stringToMimeType = (str: any): MimeType => str;
+    const format = stringToMimeType(originalBlob.type);
+
+    return { rgba: imgData, width: img.width, height: img.height, format: format };
+  } catch (e) {
+    console.error(e);
+    throw new Error(e);
+  } finally {
+    imgCanvas.remove();
+    if (imgObjectUrl) {
+      window.URL.revokeObjectURL(imgObjectUrl);
+    }
+  }
+}
+
+export async function getBlob(imageProperties: ImageProperties): Promise<Blob> {
+  const { rgba, width, height, format } = imageProperties;
+  const imgCanvas = document.createElement('canvas');
+  try {
+    imgCanvas.width = width;
+    imgCanvas.height = height;
+    const imgCtx = imgCanvas.getContext('2d');
+    const newImg = new ImageData(rgba, width, height);
+    imgCtx.putImageData(newImg, 0, 0);
+    const blob: Blob = await canvasToBlob(imgCanvas, format);
+    return blob;
+  } catch (e) {
+    console.error(e);
+    throw new Error(e);
+  } finally {
+    imgCanvas.remove();
+  }
 }
